Extract MetricsTooltip into a standalone component
