+++ ui/packages/shared/profile/src/MetricsGraph/index.tsx
@@ -11,24 +11,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, {useEffect, useRef, useState} from 'react';
+import React, {useRef, useState} from 'react';
 import * as d3 from 'd3';
 import {pointer} from 'd3-selection';
-import {formatForTimespan} from '@parca/functions/time';
-import {SingleProfileSelection, timeFormat} from '..';
-import {cutToMaxStringLength} from '@parca/functions/string';
 import throttle from 'lodash.throttle';
-import {MetricsSeries as MetricsSeriesPb, MetricsSample, Label} from '@parca/client';
-import {usePopper} from 'react-popper';
-import type {VirtualElement} from '@popperjs/core';
-import {valueFormatter, formatDate, sanitizeHighlightedValues} from '@parca/functions';
+
 import {DateTimeRange, useKeyDown} from '@parca/components';
 import {useContainerDimensions} from '@parca/dynamicsize';
+import {formatForTimespan} from '@parca/functions/time';
+import {MetricsSeries as MetricsSeriesPb, MetricsSample, Label} from '@parca/client';
+import {valueFormatter, formatDate, sanitizeHighlightedValues} from '@parca/functions';
 
 import MetricsSeries from '../MetricsSeries';
 import MetricsCircle from '../MetricsCircle';
+import MetricsTooltip from './MetricsTooltip';
+import {SingleProfileSelection} from '..';
 
-interface RawMetricsGraphProps {
+interface Props {
   data: MetricsSeriesPb[];
   from: number;
   to: number;
@@ -40,7 +39,7 @@ interface RawMetricsGraphProps {
   width?: number;
 }
 
-interface HighlightedSeries {
+export interface HighlightedSeries {
   seriesIndex: number;
   labels: Label[];
   timestamp: number;
@@ -63,7 +62,7 @@ const MetricsGraph = ({
   onLabelClick,
   setTimeRange,
   sampleUnit,
-}: RawMetricsGraphProps): JSX.Element => {
+}: Props): JSX.Element => {
   const {ref, dimensions} = useContainerDimensions();
 
   return (
@@ -95,140 +94,6 @@ export const parseValue = (value: string): number | null => {
 const lineStroke = '1px';
 const lineStrokeHover = '2px';
 
-interface MetricsTooltipProps {
-  x: number;
-  y: number;
-  highlighted: HighlightedSeries;
-  onLabelClick: (labelName: string, labelValue: string) => void;
-  contextElement: Element | null;
-  sampleUnit: string;
-}
-
-function generateGetBoundingClientRect(contextElement: Element, x = 0, y = 0): () => DOMRect {
-  const domRect = contextElement.getBoundingClientRect();
-  return () =>
-    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-    ({
-      width: 0,
-      height: 0,
-      top: domRect.y + y,
-      left: domRect.x + x,
-      right: domRect.x + x,
-      bottom: domRect.y + y,
-    } as DOMRect);
-}
-
-const virtualElement: VirtualElement = {
-  getBoundingClientRect: () => {
-    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-    return {
-      width: 0,
-      height: 0,
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-    } as DOMRect;
-  },
-};
-
-export const MetricsTooltip = ({
-  x,
-  y,
-  highlighted,
-  onLabelClick,
-  contextElement,
-  sampleUnit,
-}: MetricsTooltipProps): JSX.Element => {
-  const [popperElement, setPopperElement] = useState<HTMLDivElement | null>(null);
-
-  const {styles, attributes, ...popperProps} = usePopper(virtualElement, popperElement, {
-    placement: 'auto-start',
-    strategy: 'absolute',
-    modifiers: [
-      {
-        name: 'preventOverflow',
-        options: {
-          tether: false,
-          altAxis: true,
-        },
-      },
-      {
-        name: 'offset',
-        options: {
-          offset: [30, 30],
-        },
-      },
-    ],
-  });
-
-  const update = popperProps.update;
-
-  useEffect(() => {
-    if (contextElement != null) {
-      virtualElement.getBoundingClientRect = generateGetBoundingClientRect(contextElement, x, y);
-      void update?.();
-    }
-  }, [x, y, contextElement, update]);
-
-  const nameLabel: Label | undefined = highlighted?.labels.find(e => e.name === '__name__');
-  const highlightedNameLabel: Label = nameLabel !== undefined ? nameLabel : {name: '', value: ''};
-
-  return (
-    <div ref={setPopperElement} style={styles.popper} {...attributes.popper} className="z-10">
-      <div className="flex max-w-md">
-        <div className="m-auto">
-          <div
-            className="border-gray-300 dark:border-gray-500 bg-gray-50 dark:bg-gray-900 rounded-lg p-3 shadow-lg opacity-90"
-            style={{borderWidth: 1}}
-          >
-            <div className="flex flex-row">
-              <div className="ml-2 mr-6">
-                <span className="font-semibold">{highlightedNameLabel.value}</span>
-                <span className="block text-gray-700 dark:text-gray-300 my-2">
-                  <table className="table-auto">
-                    <tbody>
-                      <tr>
-                        <td className="w-1/4">Value</td>
-                        <td className="w-3/4">
-                          {valueFormatter(highlighted.value, sampleUnit, 1)}
-                        </td>
-                      </tr>
-                      <tr>
-                        <td className="w-1/4">At</td>
-                        <td className="w-3/4">{formatDate(highlighted.timestamp, timeFormat)}</td>
-                      </tr>
-                    </tbody>
-                  </table>
-                </span>
-                <span className="block text-gray-500 my-2">
-                  {highlighted.labels
-                    .filter((label: Label) => label.name !== '__name__')
-                    .map(function (label: Label) {
-                      return (
-                        <button
-                          key={label.name}
-                          type="button"
-                          className="inline-block rounded-lg text-gray-700 bg-gray-200 dark:bg-gray-700 dark:text-gray-400 px-2 py-1 text-xs font-bold mr-3"
-                          onClick={() => onLabelClick(label.name, label.value)}
-                        >
-                          {cutToMaxStringLength(`${label.name}="${label.value}"`, 37)}
-                        </button>
-                      );
-                    })}
-                </span>
-                <span className="block text-gray-500 text-xs">
-                  Hold shift and click label to add to query.
-                </span>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
 export const RawMetricsGraph = ({
   data,
   from,
@@ -239,7 +104,7 @@ export const RawMetricsGraph = ({
   setTimeRange,
   width,
   sampleUnit,
-}: RawMetricsGraphProps): JSX.Element => {
+}: Props): JSX.Element => {
   const graph = useRef(null);
   const [dragging, setDragging] = useState(false);
   const [hovering, setHovering] = useState(false);
